% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods_multiparm.R
\name{dyn_T}
\alias{dyn_T}
\title{Calculate calculate log T for given host and deme}
\usage{
dyn_T(
  ttree,
  obs,
  grid,
  omega,
  omega_bar,
  pit,
  off.r,
  off.p,
  pi,
  ws.shape,
  ws.scale,
  obs.start,
  obs.end,
  grid.delta = 1/365,
  host,
  deme
)
}
\arguments{
\item{ttree}{Transmission tree}

\item{obs}{Observation times}

\item{grid}{Discrete grid over which to evaluate functions}

\item{omega}{Exclusion probabilities}

\item{omega_bar}{Offspring exclusion probabilities}

\item{pit}{Observation probabilities}

\item{off.r}{Shape parameter for the number of offspring}

\item{off.p}{Probability parameter for the number of offspring}

\item{pi}{Probability of host being sampled}

\item{ws.shape}{Shape parameter of primary sampling time distribution}

\item{ws.scale}{Scale parameter of primary sampling time distribution}

\item{obs.start}{Start time of outbreak sampling}

\item{obs.end}{Stop time of outbreak sampling}

\item{grid.delta}{Discrete time step}

\item{deme}{Deme for which likelihood is calculated}

\item{hosts}{Host for which likelihood is calculated}
}
\description{
Calculate calculate log T for given host and deme
}
