% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods_multiparm.R
\name{dyn_U}
\alias{dyn_U}
\title{Calculate calculate log U for given hosts and demes}
\usage{
dyn_U(
  ttree,
  grid,
  omega,
  w.shape,
  w.scale,
  obs.end,
  grid.delta = 1/365,
  host1,
  deme1,
  host2,
  deme2,
  pm
)
}
\arguments{
\item{ttree}{Transmission tree}

\item{grid}{Discrete grid over which to evaluate functions}

\item{omega}{Exclusion probabilities}

\item{w.shape}{Shape parameter of generation time distribution}

\item{w.scale}{Scale parameter of generation time distribution}

\item{obs.end}{Stop time of outbreak sampling}

\item{grid.delta}{Discrete time step}

\item{host1}{Infector for which likelihood is calculated}

\item{deme1}{Infector deme for which likelihood is calculated}

\item{host2}{Infected for which likelihood is calculated}

\item{deme2}{Infected deme for which likelihood is calculated}

\item{pm}{Transition probability matrix between demes}
}
\description{
Calculate calculate log U for given hosts and demes
}
