% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractions.R
\name{extractParmTrace}
\alias{extractParmTrace}
\title{Extracts full parameter trace plots from a resTransPhylo object}
\usage{
extractParmTrace(record)
}
\arguments{
\item{record}{resTransPhylo object}
}
\description{
Extracts full parameter trace plots from a resTransPhylo object
}
