% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference2.R
\name{inferTTree2}
\alias{inferTTree2}
\title{Infer transmission tree given a phylogenetic tree}
\usage{
inferTTree2(
  ptree,
  w.shape = 2,
  w.scale = 1,
  ws.shape = NA,
  ws.scale = NA,
  w.mean = NA,
  w.std = NA,
  ws.mean = NA,
  ws.std = NA,
  mcmcIterations = 12000,
  thinning = 1,
  mcmc.tree.updates = NA,
  init.r = 2,
  init.p = 0.5,
  init.pi = 0.5,
  init.kappa = 0.5,
  init.lambda = 0.5,
  init.rho = 0.8,
  init.ctree = NA,
  update.r = T,
  update.p = F,
  update.pi = T,
  update.kappa = T,
  update.lambda = T,
  update.rho = F,
  update.ctree = T,
  ndemes = NA,
  r.shape = 1,
  r.scale = 1,
  p.shape1 = 1,
  p.shape2 = 1,
  pi.shape1 = 1,
  pi.shape2 = 1,
  kappa.shape = 1,
  kappa.scale = 1,
  lambda.shape = 1,
  lambda.scale = 1,
  rho.shape1 = 1,
  rho.shape2 = 1,
  demes.prior = NA,
  dateS = -Inf,
  dateT = NA,
  grid.delta = NA,
  rho.demes = NA,
  verbose = F
)
}
\arguments{
\item{ptree}{Phylogenetic tree}

\item{w.shape}{Shape parameter of the Gamma distribution representing the generation time}

\item{w.scale}{Scale parameter of the Gamma distribution representing the generation time}

\item{ws.shape}{Shape parameter of the Gamma distribution representing the sampling time}

\item{ws.scale}{Scale parameter of the Gamma distribution representing the sampling time}

\item{w.mean}{Mean of the Gamma distribution representing the generation time}

\item{w.std}{Std of the Gamma distribution representing the generation time}

\item{ws.mean}{Mean of the Gamma distribution representing the sampling time}

\item{ws.std}{Std of the Gamma distribution representing the sampling time}

\item{mcmcIterations}{Number of MCMC iterations to run the algorithm for}

\item{thinning}{MCMC thinning interval between two sampled iterations}

\item{mcmc.tree.updates}{Number of transmission tree updates per parameter update}

\item{init.r}{Starting value of offspring distribution parameter r}

\item{init.p}{Starting value of offspring distribution parameter p}

\item{init.pi}{Starting value of sampling proportion pi}

\item{init.kappa}{Starting value for the initial within-host population size}

\item{init.lambda}{Starting value for the within-host population growth rate}

\item{init.rho}{Starting value for the probability of transmission to the same location}

\item{init.ctree}{Starting value for the combined tree}

\item{update.r}{Whether to update the parameter r}

\item{update.p}{Whether to update the parameter p}

\item{update.pi}{Whether to update the parameter pi}

\item{update.kappa}{Whether to update the parameter kappa}

\item{update.lambda}{Whether to update the parameter lambda}

\item{update.rho}{Whether to update parameter rho}

\item{update.ctree}{Whether to update the transmission tree}

\item{ndemes}{Number of possible demes}

\item{r.shape}{Shape parameter for the Gamma prior of parameter r}

\item{r.scale}{Scale parameter for the Gamma prior of parameter r}

\item{p.shape1}{Shape1 parameter for the Beta prior of parameter p}

\item{p.shape2}{Shape2 parameter for the Beta prior of parameter p}

\item{pi.shape1}{Shape1 parameter for the Beta prior of parameter pi}

\item{pi.shape2}{Shape2 parameter for the Beta prior of parameter pi}

\item{kappa.shape}{Shape parameter for the Gamma prior of parameter kappa}

\item{kappa.scale}{Scale parameter for the Gamma prior of parameter kappa}

\item{lambda.shape}{Shape parameter for the Gamma prior of parameter lambda}

\item{lambda.scale}{Scale parameter for the Gamma prior of parameter lambda}

\item{rho.shape1}{Shape1 parameter for the Beta prior of parameter rho}

\item{rho.shape2}{Shape2 parameter for the Beta prior of parameter rho}

\item{demes.prior}{Prior probability for the deme of the root host}

\item{dateS}{Start date for observations}

\item{dateT}{End date for observations}

\item{grid.delta}{Grid resolution for approximating exclusion probabilities}

\item{verbose}{Whether or not to use verbose mode (default is false)}
}
\description{
Infer transmission tree given a phylogenetic tree
}
