% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_cont.R
\name{inferTTreecont}
\alias{inferTTreecont}
\title{Continues an MCMC chain from a previous set of results}
\usage{
inferTTreecont(prevrun, mcmcIterations = 12000, thinning = 1, verbose = F)
}
\arguments{
\item{prevrun}{resTransPhylo object}

\item{mcmcIterations}{Number of new MCMC iterations to perform}

\item{thinning}{MCMC thinning interval between two sampled iterations}

\item{verbose}{Whether or not to use verbose mode (default is false)}
}
\description{
Continues an MCMC chain from a previous set of results
}
