% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_lik_locs_felsenstein}
\alias{log_lik_locs_felsenstein}
\title{Demes likelihood for a transmission tree using Felsenstein pruning}
\usage{
log_lik_locs_felsenstein(ttree, pm, demes.prior)
}
\arguments{
\item{ttree}{Transmission tree}

\item{pm}{Probability transition matrix between demes}

\item{demes.prior}{Prior probability for the location of the root host}
}
\description{
Demes likelihood for a transmission tree using Felsenstein pruning
}
