% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_likelihood_coalescence_linear_cpp}
\alias{log_likelihood_coalescence_linear_cpp}
\title{C++ version of log_likelihood_coalescence_linear}
\usage{
log_likelihood_coalescence_linear_cpp(
  infected_time,
  final_time,
  start_time,
  kappa,
  lambda,
  branch_combs,
  coalescence
)
}
\arguments{
\item{infected_time}{double}

\item{final_time}{double}

\item{start_time}{double}

\item{kappa}{double}

\item{lambda}{double}

\item{branch_combs}{double}

\item{coalescence}{int (1 if event is a coalescent event, else 0)}
}
\value{
double log-likelihood increment
}
\description{
C++ version of log_likelihood_coalescence_linear
}
