% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_conversions.R
\name{ptreeFromPhylo}
\alias{ptreeFromPhylo}
\title{Converts an ape phylo object into a phylogenetic tree with host information}
\usage{
ptreeFromPhylo(tr, dateLastSample, host, demes)
}
\arguments{
\item{tr}{phylo object}

\item{dateLastSample}{date of the last sample}

\item{host}{Optional host information, if not present the leaf names needs to contain this, eg 1.1 etc}

\item{demes}{Optional location information, demes must be categorized as contiguous integers. NA can be used for missing data.}
}
\value{
phylogenetic tree
}
\description{
Converts an ape phylo object into a phylogenetic tree with host information
}
