% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ongoing.R
\name{simulateOutbreak}
\alias{simulateOutbreak}
\title{Simulate an ongoing outbreak}
\usage{
simulateOutbreak(
  off.r = 1,
  off.p = 0.5,
  kappa = 0.2,
  lambda = 0.2,
  pi = 0.5,
  sec.p = 0,
  sec.n = 0,
  sec.t = 14/365,
  w.shape = 2,
  w.scale = 1,
  ws.shape = NULL,
  ws.scale = NULL,
  w.mean = NULL,
  w.std = NULL,
  ws.mean = NULL,
  ws.std = NULL,
  dateStartOutbreak = 2000,
  dateS = NULL,
  dateT = 2010,
  grid.delta = 1/365,
  nSampled = NA,
  limit = "hosts"
)
}
\arguments{
\item{off.r}{Shape parameter for the number of offspring}

\item{off.p}{Probability parameter for the number of offspring}

\item{kappa}{Initial pathogen population}

\item{lambda}{Pathogen population linear growth rate}

\item{pi}{Probability of host being sampled}

\item{sec.p}{Probability of repeated samples if limit is "samples" or "time".}

\item{sec.n}{Number of repeated sampled if limit is "hosts".}

\item{sec.t}{Time between repeated samples}

\item{w.shape}{Shape parameter of generation time distribution}

\item{w.scale}{Scale parameter of generation time distribution}

\item{ws.shape}{Shape parameter of primary sampling time distribution}

\item{ws.scale}{Scale parameter of primary sampling time distribution}

\item{w.mean}{Mean parameter of generation time distribution}

\item{w.std}{Standard deviation of generation time distribution}

\item{ws.mean}{Mean parameter of primary sampling time distribution}

\item{ws.std}{Standard deviation of primary sampling time distribution}

\item{dateStartOutbreak}{Outbreak start date}

\item{dateS}{Start time of outbreak sampling}

\item{dateT}{Stop time of outbreak sampling}

\item{grid.delta}{Discrete time step}

\item{nSampled}{Number of sampled hosts if limit is "hosts", or number of samples if limit is "samples".}

\item{limit}{Limiting factor for the simulation: "hosts", "samples" or "time".}
}
\description{
Simulate an ongoing outbreak
}
